work in progress (add scheduler)
