+++ discord_bouncer/discord_actions.py
@@ -1,12 +1,10 @@
-import discord
-from dotenv import load_dotenv
-import os
-from .checkout_session import get_payment_link, PAID_ROLE
-from .database import user_has_paid
 import logging
-from google.cloud import secretmanager
 
-logging.basicConfig(level=logging.INFO)
+import discord
+
+from .checkout_session import PAID_ROLE, get_payment_link
+from .scheduler import update_users, user_has_paid
+from .database import get_expired_users, user_has_access
 
 logger = logging.getLogger(__name__)
 
@@ -82,31 +80,10 @@ async def add_role(
     await channel.send(f"Successfully gave the role {role.name} to {member.name}")
 
 
-async def remove_role(
+async def remove_roles(
     channel: discord.Message.channel, member: discord.Member, role: discord.Role
 ):
-    await member.remove_roles(role)
-    await channel.send(f"Successfully removed the role {role.name} from {member.name}")
-
-
-def load_secrets_into_env():
-    project_id = os.getenv("GOOGLE_CLOUD_PROJECT")
-
-    secrets_manager = secretmanager.SecretManagerServiceClient()
-    secrets = secrets_manager.list_secrets(request={"parent": f"projects/{project_id}"})
-
-    for secret in secrets:
-        secret_version = secrets_manager.access_secret_version(
-            request={"name": f"{secret.name}/versions/latest"}
-        )
-
-        secret_value = secret_version.payload.data.decode("UTF-8")
-        secret_name = secret.name.split("/")[-1]
-        logger.info(f"Setting {secret_name} in environment!")
-        os.environ[secret_name] = secret_value
-
-
-def main():
-    # load_secrets_into_env()
-    load_dotenv(override=True)
-    bot.run(os.getenv("DISCORD_KEY"))
+    #remove multiple roles
+    pass 
+    # await member.remove_roles(role)
+    # await channel.send(f"Successfully removed the role {role.name} from {member.name}")
