+++ pyproject.toml
@@ -15,6 +15,7 @@ pytz = "^2024.1"
 py-cord = "^2.5.0"
 google-cloud-firestore = "^2.16.0"
 google-cloud-secret-manager = "^2.20.0"
+apscheduler = "^3.10.4"
 
 [tool.poetry.scripts]
 bouncer = "discord_bouncer.discord_bouncer:main"
